Rearranged files to make testing possible.
