+++ tests/index.lua
@@ -1,0 +1,1 @@
+require("PlanetsLib_recipe_productivity_effects")
